Implement native Apple Pay button
For now, the implemented native button should only be used when the strange bug in the 'pay ' lib happens.

The bug  hides the Apple Pay button when there is no card available in the user's wallet or no supported cards.

This bug sometimes causes a rejection from Apple Store's review process.

Some enhancements before a full transition to the native button approach:
- Handling cancelled behavior by the user
- UI styling like the supported in the 'pay' lib
- Loading indicator